% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umart_main.R
\name{generate_data}
\alias{generate_data}
\title{Generate random data for a UMAP}
\usage{
generate_data(seed = 520, type = "norm", number = 50, ...)
}
\arguments{
\item{seed}{Seed for reproducibility}

\item{type}{Type of randomness to use, defaults to "norm" to generate data
from a normal distribution}

\item{number}{Number of samples to generate}

\item{...}{Arguments to other functions.}
}
\value{
A vector of the generated data, as requested.
}
\description{
Generate random data for a UMAP
}
\examples{
generate_data(seed = 520, type ="norm", number = 50, mean = 233, sd = 0.5)
}
