% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umart_main.R
\name{generate_umart}
\alias{generate_umart}
\title{Generate umart}
\usage{
generate_umart(umap_df, opacity = 0.25, group_col = NULL)
}
\arguments{
\item{umap_df}{An object from run_umap}

\item{opacity}{controls background opacity}

\item{group_col}{Column to use to color UMAP result}
}
\value{
UMAP art (umart)
}
\description{
Generate umart
}
\examples{
r_x <- generate_data(seed = 520, type ="norm", number = 50, mean = 233,
sd = 0.5)
r_y <- generate_data(seed = 520, type ="norm", number = 50, mean = 520,
sd = 0.05)
r_df <- data.frame(x = r_x, y = r_y, group=rep(paste0("grp_", 1:5), 10))
umap_df <- run_umap(r_df[,-3], neighbors = 20, seed = 520)
generate_umart(umap_df, group_col = r_df$group)
}
