% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umart_main.R
\name{run_umap}
\alias{run_umap}
\title{Run a UMAP}
\usage{
run_umap(df = NULL, neighbors = NULL, seed = 520, ...)
}
\arguments{
\item{df}{A data.frame/matrix with all numeric values}

\item{neighbors}{Number of neighbors to use}

\item{seed}{Seed for reproducibility}

\item{...}{Other arguments to "umap::uwot".}
}
\value{
A UMAP object
}
\description{
Run a UMAP
}
\examples{
\dontrun{
r_x <- generate_data(seed = 520, type ="norm", number = 50, mean = 233,
sd = 0.5)
r_y <- generate_data(seed = 520, type ="norm", number = 50, mean = 520,
sd = 0.05)
r_df <- data.frame(x = r_x, y = r_y)
run_umap(r_df, neighbors = 20, seed = 520)
}
}
